requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
rollbar==1.2.0
sqlparse==0.5.3
urllib3==2.3.0
